Add files via upload
Scenario 

DEF company has decided to implement a system that will provide a travel allowance to employees based on certain criteria.

Assume that you have been chosen as the developer for this system. According to the customer, the following tasks and conditions need to be fulfilled by the software:

The user should be able to enter Employee ID (empID), Employee Name (empName), Employee Salary (empSalary), and distance from office in kilometres to Employee’s home (empHomeKM).
Need to display relevant messages when taking user input.
If the employee’s home is more than a certain distance from the office, they will be eligible for a travel allowance.
The travel allowance will be a certain percentage of the salary(empSalary) of the employee.
If the distance from home is more than the certain distance, the final salary with travel allowance should be calculated.
     o   Final Salary with travel allowance (finalSal) = Employee Salary (empSalary) +  Employee Salary (empSalary) * the certain percentage

The final salary should be displayed with Employee ID and Employee Name.
important:
The certain distance should be the first two digits of your “Registration Number”.
The certain percentage should be the last two digits of your “Student Number(SNumber)”.

Include your “registration number”, “student number” as comments at the top of your code for identification.

Task

Develop a simple JAVA program which can execute above functionalities. (When creating identifiers use the names which are given in the brackets.)